+++ python/mujoco/specs_test.py
@@ -484,6 +484,14 @@ class SpecsTest(absltest.TestCase):
     # Check that the state is the same.
     np.testing.assert_array_equal(state1, state2)
 
+  def test_make_mesh(self):
+    spec = mujoco.MjSpec()
+    mesh = spec.add_mesh(name='wedge')
+    mesh.make_wedge(resolution=[25, 25], radius=1, fov=[90, 45], gamma=0)
+    model = spec.compile()
+    self.assertEqual(model.nmesh, 1)
+    self.assertEqual(model.nmeshvert, 25 * 25)
+
   def test_compile_errors_with_line_info(self):
     spec = mujoco.MjSpec()