+++ unity/Runtime/Bindings/MjBindings.cs
@@ -496,6 +496,16 @@ public enum mjtMeshInertia : int{
   mjMESH_INERTIA_LEGACY = 2,
   mjMESH_INERTIA_SHELL = 3,
 }
+public enum mjtMeshBuiltin : int{
+  mjMESH_BUILTIN_NONE = 0,
+  mjMESH_BUILTIN_SPHERE = 1,
+  mjMESH_BUILTIN_HEMISPHERE = 2,
+  mjMESH_BUILTIN_PRISM = 3,
+  mjMESH_BUILTIN_CONE = 4,
+  mjMESH_BUILTIN_TORUS = 5,
+  mjMESH_BUILTIN_WEDGE = 6,
+  mjMESH_BUILTIN_PLATE = 7,
+}
 public enum mjtBuiltin : int{
   mjBUILTIN_NONE = 0,
   mjBUILTIN_GRADIENT = 1,