+++ doc/changelog.rst
@@ -15,6 +15,7 @@ General
 - Removed the SdfLib plugin and the dependency on `SdfLib <https://github.com/UPC-ViRVIG/SdfLib>`__. SDFs are now
   supported natively in mjModel.
 - Removed ``oct_depth`` from :ref:`mjvOption` (unused).
+- Added the functionality to create a builtin meshes, see :ref:`mesh/builtin<asset-mesh-builtin>`.
 - Inertia computation in MuJoCo C is now performed by a new :ref:`pipeline<piStages>` function :ref:`mj_makeM`, which
   combines the Composite Rigid Body algorithm in :ref:`mj_crb` and additional terms related to
   :ref:`tendon armature<tendon-spatial-armature>`. Code that uses :ref:`mj_crb` to compute the inertia should now use